+++ projects/ng-esri-map/src/lib/registry/actions.ts
@@ -1,1 +1,0 @@
-export const SELECT_POINT = 'select-point';