+++ projects/ng-esri-map/src/lib/registry/index.ts
@@ -1,1 +1,1 @@
-export * from './actions';
+export { loadOptions } from './load-options';