+++ src/app/map/map.component.ts
@@ -9,13 +9,26 @@ import { NgEsriMapComponent } from 'ng-esri-map';
 export class MapComponent {
   @ViewChild('map') public map: NgEsriMapComponent;
   @Input() public set options(options) {
-    this.init(options);
+    const {scaleBar, homeButton} = this.prevOptions;
+
+    if (scaleBar === options.scaleBar && homeButton === options.homeButton) {
+      this.updateMainPoint(options);
+
+      return;
+    }
+
+    this.buildMap(options);
+
+    this.prevOptions = options;
+  }
+  @Input() public set layers(layers) {
+    this.rebuildAllLayers(layers);
   }
   @Output() public pointSelected: EventEmitter<any> = new EventEmitter();
 
-  public async init({latitude, longitude, layers, scaleBar, homeButton}) {
-    await this.rebuildAllLayers(layers);
+  private prevOptions: any = {};
 
+  public async buildMap({latitude, longitude, scaleBar, homeButton}) {
     await this.map.initMap({
       latitude,
       longitude,
@@ -26,6 +39,13 @@ export class MapComponent {
       homeButton
     });
 
+    this.updateMainPoint({latitude, longitude});
+    this.createActions();
+
+    this.map.onRightClick = event => this.showPopup(event.mapPoint);
+  }
+
+  private async updateMainPoint({latitude, longitude}) {
     await this.map.setMainPoint({
       latitude,
       longitude,
@@ -36,13 +56,11 @@ export class MapComponent {
       }
     });
 
-    this.createActions();
-
-    this.map.onRightClick = event => this.showPopup(event.mapPoint);
+    this.map.mapView.goTo(this.map.mainGraphic);
   }
 
-  private async rebuildAllLayers(layers) {
-    await this.map.clearAllLayers();
+  private rebuildAllLayers(layers) {
+    this.map.clearAllLayers();
 
     this.map.buildMapImageLayersList(layers);
   }